last update for windows(for now)
tried a lot but due to my lack of knowledge in rest api couldnot do anything. In future I will definetly do something for windows but for now I will be using web app for pc and android will be made. Started working on the project screen, will be completed by next update.