v0 completed
very sorry for the delay, was stuck in many things. Completed the base version. I know the code is very dirty but I am not cleaning this up for now since I am moving to a new tutorial after a long time. May come back after a while, but for now this is the version 0 and I might just add some bug fixes.